Add initial menu system
